+++ Rocksolid_Light/rslight/scripts/nocem.php
@@ -108,10 +108,12 @@ function delete_message($messageid, $group) {
  }
   if($CONFIG['article_database'] == '1') {
     $database = $spooldir.'/'.$group.'-articles.db3';
-    $articles_dbh = article_db_open($database);
-    $articles_query = $articles_dbh->prepare('DELETE FROM articles WHERE msgid=:messageid');
-    $articles_query->execute(['messageid' => $messageid]);
-    $articles_dbh = null;
+    if(is_file($database)) {
+      $articles_dbh = article_db_open($database);
+      $articles_query = $articles_dbh->prepare('DELETE FROM articles WHERE msgid=:messageid');
+      $articles_query->execute(['messageid' => $messageid]);
+      $articles_dbh = null;
+    }
   }
   $this_overview=$spooldir.'/'.$group.'-overview';
   if(false === (is_file($this_overview))) {
